Extract Checkbox from CheckboxWithLabel
