+++ front/app/component-library/index.tsx
@@ -22,7 +22,7 @@ import Button, {
   ButtonStyles,
 } from './components/Button';
 import CardButton from './components/CardButton';
-import Checkbox from './components/Checkbox';
+import CheckboxWithLabel from './components/CheckboxWithLabel';
 import ColorPickerInput, {
   Props as ColorPickerInputProps,
 } from './components/ColorPickerInput';
@@ -177,7 +177,7 @@ export {
   Td,
   Th,
   Tfoot,
-  Checkbox,
+  CheckboxWithLabel,
   Box,
   useWindowSize,
   useBreakpoint,
